refactor: 선호팀 관련 dto, repository, service, controller 수정
